Improve locking when assigning leases
This adds `SKIP LOCKED` to the `SELECT ... FOR UPDATE` query, which will
prevent workers from blocking others when looking for the same
requirements.

It also adds an explicit `BEGIN IMMEDIATE` if using SQLite, which locks
the database for the whole transaction rather than the default behaviour
of just locking for writing the update, since SQLite doesn't support
row locking.